Update gameover process and traffic render

+++ source/object/lane.hpp
@@ -28,7 +28,7 @@ private:
 public:
     Lane(TextureHolder* holder, Vec2 size, int pos, int len, float speed, int spawnClock = 120, bool isRed = false, int trafficClock = 0);
 
-    void render(Engine* engine, int playerLane);
+    void render(Engine* engine, float playerLane);
     void process();
     void gameoverProcess();
     void shift(Vec2 offset);
